python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skills-sdk"
version = "0.1.0"
description = "Skills SDK for Python"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["agent skills", "skills"]
dependencies = [
  "httpx>=0.27",
  "python-frontmatter>=1.1",
]

[dependency-groups]
dev = [
  "ruff>=0.9",
]
